+++ resources/js/v8/i18n.ts
@@ -1,0 +1,31 @@
+/**
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) 2017-2018 Tobias Reich
+ * Copyright (c) 2018-2026 LycheeOrg.
+ */
+
+// Populated by laravel-vue-i18n's Vite plugin (see vite.config.ts): it writes one
+// `php_{locale}.json` per `lang/{locale}/` folder at build time - the same folders
+// the backend's Settings::getLanguages reads via `scandir(lang/)` for `supported_locale`
+// (see config/app.php). Deriving `availableLocales` from this glob instead of a network
+// call keeps it in sync with the actually-bundled translations for free.
+const langs = import.meta.glob("../../../lang/*.json");
+
+const PHP_LANG_FILE = /^\.\.\/\.\.\/\.\.\/lang\/php_(.+)\.json$/;
+
+export const availableLocales: string[] = Object.keys(langs)
+	.map((path) => path.match(PHP_LANG_FILE)?.[1])
+	.filter((locale): locale is string => locale !== undefined)
+	.sort();
+
+/**
+ * `resolve` callback for the `i18nVue` plugin: lazily loads the JSON bundle for `lang`,
+ * generated by the same Vite plugin as `availableLocales` above.
+ */
+export async function resolveLocale(lang: string) {
+	const loader = langs[`../../../lang/php_${lang}.json`];
+	if (!loader) {
+		throw new Error(`Missing locale: ${lang}`);
+	}
+	return await loader();
+}
